CompanyModel was updated - now we can idenify Company by Id instead of INN. CompanyRepository was added.

+++ pimonova_WebAPI/Models/ObjectOfNEI.cs
@@ -11,7 +11,7 @@ namespace pimonova_WebAPI.Models
         public int ObjectOfNEIID { get; set; }
 
         [Required]
-        public long? CompanyID { get; set; }
+        public int? CompanyID { get; set; }
         public Company? Company { get; set; }
 
         [Required]
